+++ lib/commands/fortnite/stw/upgrade.dart
@@ -1,19 +1,28 @@
 import "package:nyxx/nyxx.dart";
 import "package:nyxx_commands/nyxx_commands.dart";
+import "package:fishstick_dart/fishstick_dart.dart";
 import "../../../../database/database_user.dart";
 import "../../../../extensions/context_extensions.dart";
 import "../../../../resources/emojis.dart";
 
-final Command skipTutorialCommand = Command(
-  "skip-tutorial",
-  "Skip your save the world gamemode tutorial.",
-  (Context ctx) async {
+final Command stwUpgradeCommand = Command(
+  "upgrade",
+  "Upgrade a save the world homebase node..",
+  (
+    Context ctx,
+    @Choices({
+      "Backpack": "HomebaseNode:skilltree_backpacksize",
+      "Storage": "HomebaseNode:skilltree_stormshieldstorage",
+    })
+    @Description("Homebase node to upgrade.")
+        String nodeId,
+  ) async {
     DatabaseUser dbUser = await ctx.dbUser;
     dbUser.fnClientSetup();
     final campaign = dbUser.fnClient.campaign;
 
     await campaign.init(dbUser.activeAccount.accountId);
-    await campaign.skipTutorial();
+    await campaign.upgradeHomebaseNode(nodeId);
 
     final EmbedBuilder embed = EmbedBuilder()
       ..author = (EmbedAuthorBuilder()
@@ -21,11 +30,11 @@ final Command skipTutorialCommand = Command(
         ..iconUrl = ctx.user.avatarURL(format: "png"))
       ..color = DiscordColor.fromHexString(dbUser.color)
       ..title =
-          "[${campaign.powerLevel.toStringAsFixed(1)}] ${dbUser.activeAccount.displayName} | Save the World Tutorial"
+          "[${campaign.powerLevel.toStringAsFixed(1)}] ${dbUser.activeAccount.displayName} | Save the World Upgrade"
       ..thumbnailUrl = dbUser.activeAccount.avatar
-      ..description = "${tick.emoji} Successfully skipped the tutorial."
-      ..timestamp = campaign.created
-      ..footer = (EmbedFooterBuilder()..text = "Account created on");
+      ..description = "${tick.emoji} Upgraded $nodeId."
+      ..timestamp = DateTime.now()
+      ..footer = (EmbedFooterBuilder()..text = client.footerText);
 
     await ctx.respond(MessageBuilder.embed(embed));
   },